refactor(files): extract shared file display helpers and simplify components
- Move typeIcons/typeIconColors/typeBgColors into fileDisplay.ts
- Extract InlineRename into its own component
- Add compareFiles helper, simplify counts and labels
