+++ src/renderer/pages/files/fileDisplay.ts
@@ -1,5 +1,7 @@
 import type { FileEntryOrigin } from '@shared/data/types/file'
 import type { FileType } from '@shared/types/file'
+import { File, FileCode, FileText, Image as ImageIcon, Music, Video } from 'lucide-react'
+import type { FC } from 'react'
 
 type FileItemCore = {
   id: string
@@ -77,4 +79,31 @@ export function formatFileSize(bytes: number | null | undefined): string {
     unitIndex += 1
   }
   return `${value >= 10 ? value.toFixed(0) : value.toFixed(1)} ${units[unitIndex]}`
+}
+
+export const typeIcons: Record<FileType, FC<{ size?: number; strokeWidth?: number; className?: string }>> = {
+  image: ImageIcon,
+  video: Video,
+  audio: Music,
+  text: FileCode,
+  document: FileText,
+  other: File
+}
+
+export const typeIconColors: Record<FileType, string> = {
+  image: 'text-pink-500/50',
+  video: 'text-violet-500/50',
+  audio: 'text-amber-500/50',
+  text: 'text-cyan-500/50',
+  document: 'text-blue-500/50',
+  other: 'text-muted-foreground/40'
+}
+
+export const typeBgColors: Record<FileType, string> = {
+  image: 'bg-pink-500/[0.04]',
+  video: 'bg-violet-500/[0.04]',
+  audio: 'bg-amber-500/[0.04]',
+  text: 'bg-cyan-500/[0.04]',
+  document: 'bg-blue-500/[0.04]',
+  other: 'bg-muted/20'
 }
